fix: issue call native on android
