+++ apps/mobile/android/app/src/main/java/com/mobile/MainApplication.kt
@@ -25,7 +25,6 @@ import com.mezon.mobile.CallStatePackage;
 import com.mezon.mobile.AudioSessionPackage;
 import com.mezon.mobile.ImageClipboardPackage;
 import com.mezon.mobile.BadgePackage;
-import com.mezon.mobile.AudioPackage;
 import io.invertase.firebase.analytics.ReactNativeFirebaseAnalyticsPackage;
 import com.swmansion.rnscreens.RNScreensPackage;
 import com.facebook.react.soloader.OpenSourceMergedSoMapping
@@ -52,7 +51,6 @@ class MainApplication : Application(), ReactApplication {
               add(RNScreensPackage())
               add(ImageClipboardPackage())
               add(BadgePackage())
-              add(AudioPackage())
             }
 
         override fun getJSMainModuleName(): String = "src/main"
